requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
autopep8==2.0.2
cffi==1.15.1
click==8.1.3
coverage==7.2.5
cryptography==41.0.0
Django==4.2.2
django-ckeditor==6.5.1
django-cors-headers==4.0.0
django-filter==23.2
django-js-asset==2.0.0
django-redis==5.2.0
django-rq==2.8.1
djangorestframework==3.14.0
djangorestframework-api-key==2.3.0
djangorestframework-simplejwt==5.2.2
flake8==6.0.0
Markdown==3.4.3
mccabe==0.7.0
Pillow==9.5.0
psycopg2-binary==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyflakes==3.0.1
PyJWT==2.7.0
pymemcache==4.0.0
python-decouple==3.8
pytz==2023.3
redis==4.5.5
rq==1.14.1
six==1.16.0
sqlparse==0.4.4
tomli==2.0.1
